requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
chromadb
tiktoken

#for installation of local packages
-e .